fixed biomass not updated bug

+++ src/main/java/com/ntros/core/world/WorldSnapshot.java
@@ -4,7 +4,7 @@ package com.ntros.core.world;
  * Immutable copy of the world state at a given tick. Produced on the sim thread, consumed on the
  * EDT — safe to read without locking. Terrain bytes are Tile ordinals.
  */
-public record WorldSnapshot(int width, int height, byte[] terrain, byte[] biomass, long tick) {
+public record WorldSnapshot(int width, int height, byte[] terrain, float[] biomass, long tick) {
 
   public static WorldSnapshot of(World world, long tick) {
     return new WorldSnapshot(
